feat: Add react-vertical-timeline-component package
Add the "react-vertical-timeline-component" package to the project dependencies. This package will be used to create a vertical timeline component in the About page. The timeline will display the work experience of the user. This addition enhances the functionality of the application by providing a visually appealing and informative timeline feature.